Create Report Card
This project is a report of a class or group in digital form. An evaluator simply put the marks or success rate of an individual in different fields. This program help evaluator to calculate total marks as well as highest marks and who got the highest marks.
I have taken an example of school in which a class teacher make a report of a class but this program can be used in other fields also like office, group etc.